pyproject.toml
[tool.poetry]
name = "data-warden"
version = "0.1.0"
description = ""
authors = ["Marian Faryna <marikfarik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.73.0"
python-rocksdb = "^0.7.0"
psycopg2 = "^2.9.3"
boto3 = "^1.21.24"
PyYAML = "^6.0"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
uvicorn = "^0.17.1"
requests = "^2.27.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
